Changed filenames; included logo images; new class
The "-inc" at the end of each include is unnecessary.

Added a few different sized logo images, 3 are Consistently Inconsistent
on one line. the other two are Consistently Inconsistent on two lines,
centered.

Removed the "Subtitle" section on each page, and replaced the
unnecessary "h4:last-of-type" selector under the Blog Page Styles with a
class styling the date called "published" which seems to be the
convention for WordPress theming.