Fixed links to fonts, CSS, imgs; moved div "row"
+ Fixed links to fonts and imgs.  Turns out the `url('../` was throwing
everything off in the WP structure.

+ In WP, the function `<?php bloginfo(); ?>` with the parameter
`template_url` (` <?php bloginfo( 'template_url' ); ?> `) pulls the
stylesheet dynamically from the main theme directory.  It turns out that
it's a requirement for a WP template to work.  Also applied to the JS
loads.

+ Commented out the fonts that are not in use in header.php.

+ Removed the ` ; ` which was functioning as a separator for the meta
viewport contents.  Should have been ` , `

+ Moved the meta group to top of the document.

+ Moved the div with `class="row"` to the header.php, to improve
legibility and simplicity of index.php which focuses on the loop.