Show Widgetized area; new unstyled classes (see comment)
The widgets work!! Check them out.  They are now editable in the WP UI
(appearance > widgets).

The h3 headers are automatically given the class of  ` widget_title `
and the divs that contain the widgets are given the class of  ` widget `
.

BOTH classes currently do not have styles.  However, we should move over
all the sidebar widget stylings to correspond with those two classes, to
provide better control in the future.